+++ app/src/main/java/com/cookandroid/week1/api/request/SignUpRequest.kt
@@ -1,4 +1,4 @@
-package com.cookandroid.week1.api
+package com.cookandroid.week1.api.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
